develop was ready for release on May 29th, but the commit never happened.
Four files were modified; this (very) commit
is the first of those four.

A fifth file, READY.FOR.RELEASE is also part
of this suite (and is just a timestamp - its
very presence a semaphore. ;)

        modified:   ainsuForth-gen-exp-m4.ino

On branch develop (most recent lineage is testing-gg)

Changes not staged for commit:

        ainsuForthsketch.cpp
        ainsuForthsketch.h
        current_target

Untracked files:

        READY.FOR.RELEASE

(originally) Wed May 29 20:34:39 UTC 2019