ELSEWHERE - THE FIXED FILE
The changes (or most of them) required for the new
Adafruit_SPIFlash lib (mid 2019).

Quite a few changes that aren't important (comments
removed en masse in several block of code).

Not incorprorated into this version - it is
a model to follow, only but the compiler will
not see it.

	new file:   getline_m4_exp.cpp___sandbox-ii_pristine_copy_

On branch exp-master-aa